Btrieve
Created new Btrieve.cs file and moved all btrieve stuff in there, except for the scan statement in functions.flow.cs
